requirements-production.txt
# ====================================
# BDC PRODUCTION REQUIREMENTS
# ====================================

# Core Flask Framework
Flask==3.0.0
Flask-CORS==6.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1

# Database & ORM
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.1
Flask-Caching==2.1.0

# Task Queue
celery==5.3.4
flower==2.0.1

# Security & Authentication
bcrypt==4.1.2
cryptography==41.0.8
PyJWT==2.8.0
itsdangerous==2.1.2

# HTTP & Networking
requests==2.31.0
urllib3==2.1.0
gunicorn==23.0.0
eventlet==0.35.2

# Data Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
WTForms==3.1.1

# File Handling & Processing
Pillow==10.3.0
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Email & Communication
python-dotenv==1.0.0
email-validator==2.1.0

# Monitoring & Logging
sentry-sdk[flask]==1.38.0
prometheus-client==0.19.0
structlog==23.2.0

# External Integrations
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.110.0
stripe==7.8.0

# AWS & Cloud Services
boto3==1.34.0
botocore==1.34.0

# Rate Limiting & Security
Flask-Limiter==3.5.0
flask-talisman==1.1.0

# Performance & Optimization
gevent==23.9.1
greenlet==3.0.2

# Utilities
click==8.1.7
python-slugify==8.0.1
Faker==20.1.0

# Testing & Development (minimal for production)
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2

# Production Server
uwsgi==2.0.23